fast_api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib
bcrypt==4.0.1
python-multipart
python-jose[cryptography]
pytest
httpx
pytest-asyncio
jinja2
aiofiles
python-dotenv